Restructure Examples, add `MultiSigExample`
