+++ shelley/chain-and-ledger/executable-spec/test/STSTests.hs
@@ -4,37 +4,23 @@
 module STSTests (stsTests) where
 
 import           Data.Either (isRight)
-import           Data.Map.Strict (Map)
-import qualified Data.Map.Strict as Map (empty, fromList, singleton)
-import qualified Data.Set as Set
-import           Data.Word (Word64)
+import qualified Data.Map.Strict as Map (empty, singleton)
 import           Test.Tasty (TestTree, testGroup)
 import           Test.Tasty.HUnit (Assertion, assertBool, testCase, (@?=))
 
-import           Cardano.Crypto.DSIGN (deriveVerKeyDSIGN, genKeyDSIGN)
-import           Crypto.Random (drgNewTest, withDRG)
-import           Examples (CHAINExample (..), ex1, ex2, ex3, ex4)
-import           MockTypes (Addr, CHAIN, KeyPair, LedgerState, MultiSig, SKey, ScriptHash, Tx,
-                     TxBody, TxId, TxIn, UTXOW, UTxOState, VKey, Wdrl)
+import           Examples (CHAINExample (..), alicePay, bobPay, carlPay, dariaPay, ex1, ex2, ex3,
+                     ex4)
+import           MockTypes (CHAIN)
+import           MultiSigExamples (aliceAndBob, aliceAndBobOrCarl, aliceAndBobOrCarlAndDaria,
+                     aliceAndBobOrCarlOrDaria, aliceOnly, aliceOrBob, applyTxWithScript, bobOnly)
 
 import           BaseTypes (Seed (..))
-import           Coin (Coin (..))
-import           Control.State.Transition (PredicateFailure, TRC (..), applySTS)
-import           Keys (pattern Dms, pattern KeyPair, pattern SKey, pattern VKey, hashKey, vKey)
-import           LedgerState (genesisId, genesisState, _utxoState)
-import           PParams (emptyPParams)
+import           Control.State.Transition (TRC (..), applySTS)
 import           Slot (Slot (..))
 import           STS.Updn (UPDN)
 import           STS.Utxow (PredicateFailure (..))
 import           Tx (hashScript)
-import           TxData (pattern AddrBase, pattern KeyHashObj, pattern RequireAllOf,
-                     pattern RequireAnyOf, pattern RequireMOf, pattern RequireSignature,
-                     pattern RewardAcnt, pattern ScriptHashObj, pattern StakeKeys,
-                     pattern StakePools, pattern Tx, pattern TxBody, pattern TxIn, pattern TxOut,
-                     _body)
-import           Updates (emptyUpdate)
-import           UTxO (makeWitnessesVKey, txid)
-
+import           TxData (pattern RewardAcnt, pattern ScriptHashObj)
 
 -- | The UPDN transition should update both the evolving nonce and
 -- the candidate nonce during the first two-thirds of the epoch.
@@ -80,7 +66,6 @@ stsTests = testGroup "STS Tests"
   , testCase "CHAIN example 2 - register stake key" testCHAINExample2
   , testCase "CHAIN example 3 - create reward update" testCHAINExample3
   , testCase "CHAIN example 4 - new epoch changes" testCHAINExample4
-  , testCase "apply Transaction to genesis UTxO" testInitialUTXO
   , testCase "Alice uses SingleSig script" testAliceSignsAlone
   , testCase "FAIL: Alice doesn't sign in multi-sig" testAliceDoesntSign
   , testCase "Everybody signs in multi-sig" testEverybodySigns
@@ -108,199 +93,6 @@ stsTests = testGroup "STS Tests"
   , testCase "withdraw from script locked account, different script" testRwdAliceSignsAlone''
   , testCase "FAIL: withdraw from script locked account, signed, missing script" testRwdAliceSignsAlone'''
   ]
-
--- Multi-Signature tests
-
-mkKeyPair :: (Word64, Word64, Word64, Word64, Word64) -> (SKey, VKey)
-mkKeyPair seed = fst . withDRG (drgNewTest seed) $ do
-  sk <- genKeyDSIGN
-  return (SKey sk, VKey $ deriveVerKeyDSIGN sk)
-
-alicePay :: KeyPair
-alicePay = KeyPair vk sk
-  where (sk, vk) = mkKeyPair (0, 0, 0, 0, 0)
-
-aliceStake :: KeyPair
-aliceStake = KeyPair vk sk
-  where (sk, vk) = mkKeyPair (1, 1, 1, 1, 1)
-
-mkAddr :: (KeyPair, KeyPair) -> Addr
-mkAddr (payKey, stakeKey) =
-  AddrBase (KeyHashObj . hashKey $ vKey payKey) (KeyHashObj . hashKey $ vKey stakeKey)
-
-aliceAddr :: Addr
-aliceAddr = mkAddr (alicePay, aliceStake)
-
-
-bobPay :: KeyPair
-bobPay = KeyPair vk sk
-  where (sk, vk) = mkKeyPair (2, 2, 2, 2, 2)
-
-bobStake :: KeyPair
-bobStake = KeyPair vk sk
-  where (sk, vk) = mkKeyPair (3, 3, 3, 3, 3)
-
-bobAddr :: Addr
-bobAddr = mkAddr (bobPay, bobStake)
-
-
-carlPay :: KeyPair
-carlPay = KeyPair vk sk
-  where (sk, vk) = mkKeyPair (4, 4, 4, 4, 4)
-
-carlStake :: KeyPair
-carlStake = KeyPair vk sk
-  where (sk, vk) = mkKeyPair (5, 5, 5, 5, 5)
-
-carlAddr :: Addr
-carlAddr = mkAddr (carlPay, carlStake)
-
-
-dariaPay :: KeyPair
-dariaPay = KeyPair vk sk
-  where (sk, vk) = mkKeyPair (6, 6, 6, 6, 6)
-
-dariaStake :: KeyPair
-dariaStake = KeyPair vk sk
-  where (sk, vk) = mkKeyPair (7, 7, 7, 7, 7)
-
-dariaAddr :: Addr
-dariaAddr = mkAddr (dariaPay, dariaStake)
-
-
--- Multi-signature scripts
-singleKeyOnly :: Addr -> MultiSig
-singleKeyOnly (AddrBase (KeyHashObj pk) _ ) = RequireSignature pk
-singleKeyOnly _ = error "use VKey address"
-
-aliceOnly :: MultiSig
-aliceOnly = singleKeyOnly aliceAddr
-
-aliceOrBob :: MultiSig
-aliceOrBob = RequireAnyOf [aliceOnly, singleKeyOnly bobAddr]
-
-aliceAndBob :: MultiSig
-aliceAndBob = RequireAllOf [aliceOnly, singleKeyOnly bobAddr]
-
-aliceAndBobOrCarl :: MultiSig
-aliceAndBobOrCarl = RequireMOf 1 [aliceAndBob, singleKeyOnly carlAddr]
-
-aliceAndBobOrCarlAndDaria :: MultiSig
-aliceAndBobOrCarlAndDaria =
-  RequireAnyOf [aliceAndBob,
-                RequireAllOf [singleKeyOnly carlAddr, singleKeyOnly dariaAddr]]
-
-aliceAndBobOrCarlOrDaria :: MultiSig
-aliceAndBobOrCarlOrDaria =
-  RequireMOf 1 [aliceAndBob,
-                RequireAnyOf [singleKeyOnly carlAddr, singleKeyOnly dariaAddr]]
-
-
-initTxBody :: [(Addr, Coin)] -> TxBody
-initTxBody addrs = TxBody
-        (Set.fromList [TxIn genesisId 0, TxIn genesisId 1])
-        (map (uncurry TxOut) addrs)
-        []
-        Map.empty
-        (Coin 0)
-        (Slot 0)
-        emptyUpdate
-
-scriptTxBody :: [TxIn] -> Addr -> Wdrl -> Coin -> TxBody
-scriptTxBody inp addr wdrl c =
-  TxBody
-    (Set.fromList inp)
-    [TxOut addr c]
-    []
-    wdrl
-    (Coin 0)
-    (Slot 10)
-    emptyUpdate
-
-makeTx :: TxBody -> [KeyPair] -> Map ScriptHash MultiSig -> Tx
-makeTx txBody keyPairs =
-  Tx txBody (makeWitnessesVKey txBody keyPairs)
-
-aliceInitCoin :: Coin
-aliceInitCoin = 10000
-
-bobInitCoin :: Coin
-bobInitCoin = 1000
-
-genesis :: LedgerState
-genesis = genesisState
-           [ TxOut aliceAddr aliceInitCoin
-           , TxOut bobAddr bobInitCoin]
-
--- | Create an initial UTxO state where Alice has 'aliceInitCoin' and Bob
--- 'bobInitCoin' to spend. Then create and apply a transaction which, if
--- 'aliceKeep' is greater than 0, gives that amount to Alice and creates outputs
--- locked by a script for each pair of script, coin value in 'msigs'.
-initialUTxOState
-  :: Coin
-  -> [(MultiSig, Coin)]
-  -> (TxId, Either [[PredicateFailure UTXOW]] UTxOState)
-initialUTxOState aliceKeep msigs =
-  let addresses =
-        [(aliceAddr, aliceKeep) | aliceKeep > 0] ++
-        map (\(msig, c) ->
-               (AddrBase
-                (ScriptHashObj $ hashScript msig)
-                (ScriptHashObj $ hashScript msig), c)) msigs
-  in
-  let tx = makeTx (initTxBody addresses)
-                  [alicePay, bobPay]
-                  Map.empty in
-  (txid $ _body tx, applySTS @UTXOW (TRC( (Slot 0
-                                           , emptyPParams
-                                           , StakeKeys Map.empty
-                                           , StakePools Map.empty
-                                           , Dms Map.empty)
-                                         , _utxoState genesis
-                                         , tx)))
-
-testInitialUTXO :: Assertion
-testInitialUTXO =
-  assertBool s (isRight utxoSt')
-  where (_, utxoSt') = initialUTxOState 0 [(aliceOnly, 11000)]
-        s = "problem: " ++ show utxoSt'
-
-
--- | Start from genesis, consume Alice's and Bob's coins, create an output
--- spendable by Alice if 'aliceKeep' is greater than 0. For each pair of script
--- and coin value in 'lockScripts' create an output of that value locked by the
--- script. Sign the transaction with keys in 'signers'. Then create an
--- transaction that uses the scripts in 'unlockScripts' (and the output for
--- 'aliceKeep' in the case of it being non-zero) to spend all funds back to
--- Alice. Return resulting UTxO state or collected errors
-applyTxWithScript
-  :: [(MultiSig, Coin)]
-  -> [MultiSig]
-  -> Wdrl
-  -> Coin
-  -> [KeyPair]
-  -> Either [[PredicateFailure UTXOW]] UTxOState
-applyTxWithScript lockScripts unlockScripts wdrl aliceKeep signers = utxoSt'
-  where (txId, initUtxo) = initialUTxOState aliceKeep lockScripts
-        utxoSt = case initUtxo of
-                   Right utxoSt'' -> utxoSt''
-                   _                      -> error ("must fail test before"
-                                                   ++ show initUtxo)
-        txbody = scriptTxBody inputs aliceAddr wdrl (aliceInitCoin + bobInitCoin + sum wdrl)
-        inputs = [TxIn txId (fromIntegral n) | n <-
-                     [0..length lockScripts - (if aliceKeep > 0 then 0 else 1)]]
-                                 -- alice? + scripts
-        tx = makeTx
-              txbody
-              signers
-              (Map.fromList $ map (\scr -> (hashScript scr, scr)) unlockScripts)
-        utxoSt' = applySTS @UTXOW (TRC( (Slot 0
-                                        , emptyPParams
-                                        , StakeKeys Map.empty
-                                        , StakePools Map.empty
-                                        , Dms Map.empty)
-                                      , utxoSt
-                                      , tx))
 
 testAliceSignsAlone :: Assertion
 testAliceSignsAlone =
@@ -478,14 +270,12 @@ testRwdAliceSignsAlone' =
   utxoSt' @?= Left [[ScriptWitnessNotValidatingUTXOW]]
   where utxoSt' =
           applyTxWithScript [(aliceOnly, 11000)] [aliceOnly, bobOnly] (Map.singleton (RewardAcnt (ScriptHashObj $ hashScript bobOnly)) 1000) 0 [alicePay]
-        bobOnly = singleKeyOnly bobAddr
 
 testRwdAliceSignsAlone'' :: Assertion
 testRwdAliceSignsAlone'' =
   assertBool s (isRight utxoSt')
   where utxoSt' =
           applyTxWithScript [(aliceOnly, 11000)] [aliceOnly, bobOnly] (Map.singleton (RewardAcnt (ScriptHashObj $ hashScript bobOnly)) 1000) 0 [alicePay, bobPay]
-        bobOnly = singleKeyOnly bobAddr
         s = "problem: " ++ show utxoSt'
 
 testRwdAliceSignsAlone''' :: Assertion
@@ -493,4 +283,3 @@ testRwdAliceSignsAlone''' =
   utxoSt' @?= Left [[MissingScriptWitnessesUTXOW]]
   where utxoSt' =
           applyTxWithScript [(aliceOnly, 11000)] [aliceOnly] (Map.singleton (RewardAcnt (ScriptHashObj $ hashScript bobOnly)) 1000) 0 [alicePay, bobPay]
-        bobOnly = singleKeyOnly bobAddr
